refactor Utils namespace to Services
