+++ app/Services/Search/SearchInterface.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Utils\Search;
+namespace App\Services\Search;
 
 /**
  * Interface SearchInterface
